UnhandledExceptionMonitor class
