+++ packages/js/src/server.ts
@@ -2,7 +2,7 @@ import os from 'os'
 import domain from 'domain'
 import { Client, Util, Types } from '@honeybadger-io/core'
 import { getSourceFile } from './server/util'
-import uncaughtException from './server/integrations/uncaught_exception'
+import uncaughtException from './server/integrations/uncaught_exception_plugin'
 import unhandledRejection from './server/integrations/unhandled_rejection'
 import { errorHandler, requestHandler } from './server/middleware'
 import { lambdaHandler } from './server/aws_lambda'
